iio-monitor: fix Argument cannot be negative
The function `getmaxyx(win, row, col);` can return negative values if
things are in a subwindow, and then these values are passed to the
function `newwin`, which requires positive values. This sets up a
potential error.

Fix this by setting a min screen size (10 rows x 50 columns) to display
content, and gracefully erroring if the size is too small.

Signed-off-by: Robin Getz <robin.getz@analog.com>